refactor: 응답 dto 생성은 정적메서드로 생성하도록 통일
