+++ src/main/java/com/neuma573/autoboard/board/service/BoardService.java
@@ -49,8 +49,6 @@ public class BoardService {
                     .map(BoardResponse::ofWithoutDeleted)
                     .collect(Collectors.toList());
         }
-
-
     }
 
     @Transactional
