+++ src/main/java/com/neuma573/autoboard/global/model/entity/BaseEntity.java
@@ -4,12 +4,12 @@ import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
-import lombok.Setter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 @MappedSuperclass
 @Getter
@@ -24,8 +24,14 @@ public class BaseEntity {
     @Column(name = "updated_at", nullable = false)
     private LocalDateTime updatedAt;
 
-    @Setter
     @Column(name = "deleted_at")
     private LocalDateTime deletedAt;
 
+    public void delete() {
+        this.deletedAt = LocalDateTime.now();
+    }
+
+    public String getFormattedCreatedAt() {
+        return getCreatedAt().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
+    }
 }
