+++ src/main/java/com/neuma573/autoboard/user/model/dto/UserResponse.java
@@ -1,5 +1,6 @@
 package com.neuma573.autoboard.user.model.dto;
 
+import com.neuma573.autoboard.user.model.entity.User;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -9,4 +10,12 @@ public class UserResponse {
     private Long id;
     private String email;
     private String name;
+
+    public static UserResponse of(User user) {
+        return UserResponse.builder()
+                .id(user.getId())
+                .email(user.getEmail())
+                .name(user.getName())
+                .build();
+    }
 }
