+++ src/main/java/com/neuma573/autoboard/user/model/entity/User.java
@@ -2,8 +2,6 @@ package com.neuma573.autoboard.user.model.entity;
 
 import com.neuma573.autoboard.global.model.entity.BaseEntity;
 import com.neuma573.autoboard.global.model.enums.Status;
-import com.neuma573.autoboard.security.model.entity.VerificationToken;
-import com.neuma573.autoboard.user.model.dto.UserResponse;
 import com.neuma573.autoboard.user.model.enums.Role;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
@@ -11,7 +9,6 @@ import lombok.*;
 
 import java.time.LocalDateTime;
 import java.util.Set;
-import java.util.UUID;
 
 @NoArgsConstructor
 @AllArgsConstructor
@@ -45,14 +42,6 @@ public class User extends BaseEntity {
     @OneToMany(mappedBy = "user",cascade = CascadeType.REMOVE, fetch = FetchType.LAZY)
     private Set<UserRole> roles;
 
-    public UserResponse toResponse() {
-        return UserResponse.builder()
-                .id(this.id)
-                .email(this.email)
-                .name(this.name)
-                .build();
-    }
-
     public void addRole(UserRole userRole){
         roles.add(userRole);
     }
@@ -73,14 +62,4 @@ public class User extends BaseEntity {
         return getRoles().stream()
                 .anyMatch(userRole -> userRole.getRole() == Role.ADMIN);
     }
-
-    public VerificationToken generateVerificationToken() {
-        return VerificationToken
-                .builder()
-                .token(UUID.randomUUID().toString())
-                .email(email)
-                .expiryDate(LocalDateTime.now().plusHours(1))
-                .build();
-    }
-
 }
